Fixed so that learning row now reports new values back to learning manager and then on to OutputManager, enabling GUI-set outputs to form training labels. Also fixed problem with learning row where Slider was overriding textbox input that was outside soft bounds.

+++ src/wekimini/osc/OSCSender.java
@@ -96,7 +96,7 @@ public class OSCSender {
             Object[] o = new Object[data.length];
             try {
                 for (int i = 0; i < data.length; i++) {
-                    o[i] = data[i];
+                    o[i] = (float)data[i];
                 }
 
                 OSCMessage msg = new OSCMessage(sendMessage, o);
